feat: `/[guild]/dashboard` page
