+++ src/components/[guild]/EditGuild/components/DeleteGuildButton/DeleteGuildButton.tsx
@@ -1,27 +1,23 @@
 import { useDisclosure } from "@chakra-ui/react"
-import DeleteButton from "components/[guild]/DeleteButton"
+import Button from "components/common/Button"
 import ConfirmationAlert from "components/create-guild/Requirements/components/ConfirmationAlert"
 import useDeleteGuild from "./hooks/useDeleteGuild"
 
-type Props = {
-  beforeDelete?: () => void
-}
-
-const DeleteGuildButton = ({ beforeDelete }: Props): JSX.Element => {
+const DeleteGuildButton = (): JSX.Element => {
   const { onSubmit, isLoading } = useDeleteGuild()
   const { isOpen, onOpen, onClose } = useDisclosure()
 
   return (
     <>
-      <DeleteButton label="Delete guild" onClick={onOpen} />
+      <Button colorScheme="red" maxW="max-content" onClick={onOpen}>
+        Delete guild
+      </Button>
+
       <ConfirmationAlert
         isLoading={isLoading}
         isOpen={isOpen}
         onClose={onClose}
-        onConfirm={() => {
-          beforeDelete?.()
-          onSubmit()
-        }}
+        onConfirm={() => onSubmit()}
         title="Delete guild"
         description="Are you sure you want to delete this guild?"
         confirmationText="Delete"
