+++ src/components/[guild]/EditGuild/hooks/useEditGuild.ts
@@ -7,7 +7,7 @@ import useSubmit from "hooks/useSubmit"
 import { useRouter } from "next/router"
 import { GuildBase } from "types"
 import replacer from "utils/guildJsonReplacer"
-import { EditGuildForm } from "../EditGuildDrawer"
+import { EditGuildForm } from "../types"
 
 type Props = {
   onSuccess?: () => void
