+++ src/components/common/FloatingFooter.tsx
@@ -1,4 +1,10 @@
-import { Box, Container, ContainerProps } from "@chakra-ui/react"
+import {
+  Box,
+  BoxProps,
+  Container,
+  ContainerProps,
+  forwardRef,
+} from "@chakra-ui/react"
 import useIsStuck from "hooks/useIsStuck"
 import { PropsWithChildren } from "react"
 import Card from "./Card"
@@ -13,41 +19,41 @@ type Props = { maxWidth?: ContainerProps["maxWidth"] }
  */
 
 // for fixed, always floating footers, like now on create-guild page
-const FloatingFooter = ({
-  maxWidth = "container.lg",
-  children,
-}: PropsWithChildren<Props>) => (
-  <Box
-    position="fixed"
-    bottom={0}
-    left={0}
-    w="full"
-    zIndex={1201} // above intercom floating button
-  >
-    <Container maxWidth={maxWidth} px={{ base: 0, md: 8, lg: 10 }}>
-      {/**
-       * Intercom: This box keeps the container padding, so the Card inside could be
-       * `width:100%`
-       */}
-      <Box position="relative">
-        <Card
-          borderRadius={0}
-          borderTopRadius={{ md: "2xl" }}
-          borderWidth={{ base: "1px 0 0 0", md: "1px 1px 0 1px" }}
-          shadow="rgba(0, 0, 0, 0.1) 0px 5px 10px,rgba(0, 0, 0, 0.2) 0px 15px 40px"
-          position="absolute"
-          bottom={0}
-          w="full"
-        >
-          {children}
-        </Card>
-      </Box>
-    </Container>
-  </Box>
+const FloatingFooter = forwardRef<BoxProps, "div">(
+  ({ maxWidth = "container.lg", children }, ref) => (
+    <Box
+      ref={ref}
+      position="fixed"
+      bottom={0}
+      left={0}
+      w="full"
+      zIndex={1201} // above intercom floating button
+    >
+      <Container maxWidth={maxWidth} px={{ base: 0, md: 8, lg: 10 }}>
+        {/**
+         * Intercom: This box keeps the container padding, so the Card inside could be
+         * `width:100%`
+         */}
+        <Box position="relative">
+          <Card
+            borderRadius={0}
+            borderTopRadius={{ md: "2xl" }}
+            borderWidth={{ base: "1px 0 0 0", md: "1px 1px 0 1px" }}
+            shadow="rgba(0, 0, 0, 0.1) 0px 5px 10px,rgba(0, 0, 0, 0.2) 0px 15px 40px"
+            position="absolute"
+            bottom={0}
+            w="full"
+          >
+            {children}
+          </Card>
+        </Box>
+      </Container>
+    </Box>
+  )
 )
 
 // for sticky footers, similar to StickyBar, but on the bottom
-export const StickyFooter = ({ children, ...rest }: PropsWithChildren<any>) => {
+const StickyFooter = ({ children, ...rest }: PropsWithChildren<any>) => {
   const { ref, isStuck } = useIsStuck()
 
   return (
@@ -71,4 +77,4 @@ export const StickyFooter = ({ children, ...rest }: PropsWithChildren<any>) => {
   )
 }
 
-export default FloatingFooter
+export { FloatingFooter, StickyFooter }
