+++ src/components/create-guild/Description.tsx
@@ -1,5 +1,5 @@
 import { FormControl, FormLabel, Textarea } from "@chakra-ui/react"
-import { EditGuildForm } from "components/[guild]/EditGuild/EditGuildDrawer"
+import { EditGuildForm } from "components/[guild]/EditGuild/types"
 import { useFormContext } from "react-hook-form"
 
 const Description = ({ isDisabled = false, ...rest }): JSX.Element => {
